Support 0.14 and drop react-tools.
This package runs fine with React 0.14 so I'm dropping the `< 0.14.0`.

As I understand it there are also others (incl. me) using the React alpha because it's the only way to get parent context.
So I'm including support for the 0.14.0-alpha3 currently in npm in the peerDependencies.
((If I don't do this, then react-cropper's use of peerDeps makes npm completely refuse to install the alpha even when you explicitly ask for it))

Also react-tools is being deprecated and it doesn't look like this module is using it anymore.